requirements.txt

streamlit
streamlit-drawable-canvas
pymupdf
pillow
